Core: include optimization
