+++ include/inviwo/core/util/stdextensions.h
@@ -33,6 +33,7 @@
 
 #include <inviwo/core/util/foreacharg.h>
 #include <inviwo/core/util/hashcombine.h>
+#include <inviwo/core/util/typetraits.h>
 
 #include <warn/push>
 #include <warn/ignore/all>
@@ -44,7 +45,6 @@
 #include <numeric>
 #include <vector>
 #include <type_traits>
-#include <future>
 #include <utility>
 #include <tuple>
 #include <warn/pop>
@@ -148,54 +148,6 @@ struct overloaded : Ts... {
 };
 template <class... Ts>
 overloaded(Ts...) -> overloaded<Ts...>;
-
-struct identity {
-    template <typename T>
-    constexpr decltype(auto) operator()(T&& t) const noexcept {
-        return std::forward<T>(t);
-    }
-};
-
-struct alwaysTrue {
-    template <typename T>
-    constexpr bool operator()(T&&) const noexcept {
-        return true;
-    }
-};
-
-struct identifier {
-    template <typename T>
-    constexpr decltype(auto) operator()(T&& item) const noexcept {
-        return item.getIdentifier();
-    }
-    template <typename T>
-    constexpr decltype(auto) operator()(T* item) const noexcept {
-        return item->getIdentifier();
-    }
-    template <typename T>
-    constexpr decltype(auto) operator()(const std::unique_ptr<T>& item) const noexcept {
-        return item->getIdentifier();
-    }
-    template <typename T>
-    constexpr decltype(auto) operator()(const std::shared_ptr<T>& item) const noexcept {
-        return item->getIdentifier();
-    }
-};
-
-// type trait to check if T is derived from std::basic_string
-namespace detail {
-template <typename T, class Enable = void>
-struct is_string : std::false_type {};
-
-template <typename T>
-struct is_string<
-    T, std::void_t<typename T::value_type, typename T::traits_type, typename T::allocator_type>>
-    : std::is_base_of<std::basic_string<typename T::value_type, typename T::traits_type,
-                                        typename T::allocator_type>,
-                      T> {};
-}  // namespace detail
-template <typename T>
-struct is_string : detail::is_string<T> {};
 
 template <typename T, typename V>
 auto erase_remove(T& cont, const V& elem)
@@ -551,12 +503,6 @@ auto table(Generator gen, int start, int end, int step = 1)
         count++;
     }
     return res;
-}
-
-template <typename T>
-bool is_future_ready(const std::future<T>& future) {
-    return (future.valid() && future.wait_for(std::chrono::duration<int, std::milli>(0)) ==
-                                  std::future_status::ready);
 }
 
 /**
