+++ resources/changelog.html
@@ -30,8 +30,11 @@
 <p>Optimized includes in <code class="notranslate">core/util</code> with some functions being moved to separate header files.<br>
 Functions moved from <code class="notranslate">core/util/stringconversion.h</code> include</p>
 <ul>
-<li><code class="notranslate">std::string parseTypeIdName()</code>, now in <code class="notranslate">namespace util</code> in <code class="notranslate">core/util/demangle.h</code></li>
-<li><code class="notranslate">std::string msToString()</code> and <code class="notranslate">std::string durationToString()</code>, now in <code class="notranslate">namespace util</code> in <code class="notranslate">core/util/chronoutils.h</code></li>
+<li><code class="notranslate">parseTypeIdName()</code>, now in <code class="notranslate">namespace util</code> in <code class="notranslate">core/util/demangle.h</code></li>
+<li><code class="notranslate">msToString()</code> and <code class="notranslate">durationToString()</code>, now in <code class="notranslate">namespace util</code> in <code class="notranslate">core/util/chronoutils.h</code><br>
+Functions and structs moved from <code class="notranslate">core/util/stdextensions.h</code> include</li>
+<li><code class="notranslate">is_future_ready()</code> moved to <code class="notranslate">core/util/stdfuture.h</code></li>
+<li><code class="notranslate">struct identifier</code>, <code class="notranslate">struct identity</code>, <code class="notranslate">struct alwaysTrue</code>, and <code class="notranslate">struct is_string</code> moved to <code class="notranslate">core/util/typetraits.h</code></li>
 </ul>
 <p>Added <code class="notranslate">core/util/glmmat.h</code> containing light-weight forward declarations of glm matrices similar to <code class="notranslate">glmvec.h</code>, which no longer includes <code class="notranslate">glm.hpp</code>.<br>
 This means that it might be necessary to include <code class="notranslate">core/util/glm.h</code> where needed.<br>
